解决bug

+++ nginx.sh
@@ -42,7 +42,7 @@ make && make install
 mkdir -p ${NGINX_PATH}/temp
 
 cp -r ../html/* ${NGINX_PATH}/html/
-rm ${NGINX_PATH}/html/index.html
+[ -f ${NGINX_PATH}/html/index.html ] && rm ${NGINX_PATH}/html/index.html
 
 cp -r ../nginxconfig.io ${NGINX_PATH}/conf
